+++ runtime/configs/config.go
@@ -120,7 +120,6 @@ func NewDefaultConfig(options ...func(*Config)) *Config {
 			},
 		},
 		Utility: &UtilityConfig{
-			// TODO: add a servicer default config
 			ServicerConfig: &ServicerConfig{
 				Chains: []string{"0001"},
 			},