+++ runtime/manager_test.go
@@ -4203,6 +4203,9 @@ func TestNewManagerFromReaders(t *testing.T) {
 						ServerModeEnabled: true,
 					},
 					Utility: &configs.UtilityConfig{
+						ServicerConfig: &configs.ServicerConfig{
+							Chains: []string{"0001"},
+						},
 						MaxMempoolTransactionBytes: 1073741824,
 						MaxMempoolTransactions:     9000,
 					},