+++ shared/modules/utility_module.go
@@ -115,3 +115,9 @@ type LeaderUtilityUnitOfWork interface {
 type ReplicaUtilityUnitOfWork interface {
 	UtilityUnitOfWork
 }
+
+type Servicer interface {
+	Module
+
+	HandleRelay(*coreTypes.Relay) (*coreTypes.RelayResponse, error)
+}